refactor: (#759) router 통일성 있게 코드 수정 및 admin/notices 폴더로 파일 이동
