+++ frontend/src/pages/admin/notices/NoticeAdminPage/style.ts
